requirements.txt
Flask==2.0.2
flask-cors==3.0.10
google-generativeai==0.1.0
gunicorn==20.1.0
werkzeug==2.2.2

